Add server struct that can shutdown on command
This is only one part in implementing server shutdown. For shutdown to
be fully realized, streams or futures need to be able to close a socket
of an incoming connection.

The included integration test in tests/test_server includes an example
where the client closes the write portion of its socket so that the
server can actually shut down.