Rename module: message -> core
The ``siminau-rpc`` crate will contain specific messages types that will
be built off of the base request, response, and notification types. In
an effort to make the naming conventions make sense, the specific
message types will live in a module named message. So as to not conflict
with this, the current message module is renamed to core.

This change makes all the core types (ie RequestMessage,
ResponseMessage, and NotificationMessage) live in the core module.