Define message codes
Message codes are used with Request, Response, and Notification messages
to make it easy to understand the scope or intent of a particular
message.

Message codes are also versioned using an API level number. This makes
it easy to add new messages. API leveling in conjunction with the
project's overall semantic version number provides a way to change the
messages and each message's arguments in a controlled and easily
communicable manner.

The current message codes define API level 1.