Implement builders for non-versioned messages
While all messages are based on the Request, Response, and
Notification structs found in the core module, these structs are more
general purpose. As a result, creating specific siminau protocol
messages is a bit tedious.

This changeset implements siminau protocol message builders. These
builders have the responsibility to ensure that they create well-formed
messages, as well as making it easier to create messages.

The increase in ergonomics is quite large. For example, without using a
builder, one would have to write a request as follows::

    use rmpv::Value;
    use message::{Request, RequestCode};

    let msgid = 42;
    let args = Value::Array(vec![Value::from(1)]);
    let request = Request::new(msgid, RequestCode::Version, args);

The above doesn't ensure that there should only be a single u32
argument.

Using a builder, the above code is reduced to this::

    use message::request;

    let req = request(42).version(1);

The above will ensure that there is only a single message argument, and
that argument is a u32 integer value.