New type: OpenMode
OpenMode is a type to help manage the mode bitfield.

Using a bitfield to represent a file's open mode is useful for keeping
size down to the bare minimum. This is helpful especially when
transmitting messages across the network, since a file's mode +
associated flags can be represented with a single byte.

OpenMode is comprised of 2 different types of information:

1. kind
2. flags

The OpenMode's kind can only have a single value, and this value
represents what how a file will be accessed.

The OpenMode's flags can be any, some, all, or none of the flags. As
such, it makes sense for the flags to be a bitfield.

Since an OpenMode's kind can only be a single value, it cannot itself be
a bitfield. Thus, the need for the OpenMode type to manage both the
single value kind as well as the potentially multiple value flags all
using the same byte.

This changeset also refactors the message::v1 module into a proper
submodule; the OpenMode type lives in message::v1::util.
test::message::v1 is also similarly refactored where ResponseBuilder,
RequestBuilder, and util each have their own dedicated modules. This was
done to help readability as the message::v1 module was starting to get
too large to keep everything in mind when reading.