RequestBuilder: refactor into separate module
As both ``RequestBuilder`` and ``ResponseBuilder`` will have many
methods, the size of the message::v1 module becomes unwieldy (in terms
of source code line count and organization) and harder to read.

In order to simplify, RequestBuilder and ResponseBuilder will be
refactored into their own sub-modules under the ``message::v1`` module.
``message::v1`` will continue to be the home for component types such as
``RequestCode``, ``ResponseCode``, ``Request``, and ``Response``.

This changeset refactors only ``RequestBuilder``.