OpenMode: new method from_bits()
OpenModeBuilder is only meant to create an OpenMode object from separate
OpenKind and OpenFlag values. However, there also needs to be a way to
create an OpenMode object directly from the bits of a u8 value. One
example of this would be if receiving a u8 value representing the mode
from over the network and needing to create an OpenMode from the u8.

This changeset also introduces 1 public and 1 private associated
constants to the OpenMode struct.

The public associated const is INVALID_BITS; this makes it easier for
external code to determine for themselves what u8 bits are considered
invalid all while not breaking if the value of OpenMode::INVALID_BITS
ever changes in the future.

The private associated const is OPENKIND_BITS and is only meant to be
used internally within the OpenMode struct's methods. This was
implemented as a way to reduce hard-coding the exact value in each
method which, if OPENKIND_BITS ever changes value in the future, will
ensure the methods still work without modification.