Builder methods for versioned messages: Create
Message builders makes it easier to create valid messages.

This changeset implements building messages for RequestCode::Create and
ResponseCode::Create. Unit/functional tests are also included.

This changeset also does a refactor for ResponseBuilder where the
``open`` and ``create`` methods actually call a private method. This is
due to the open and create response messages actually being exactly the
same except for the message code. This was done in order to uphold the
"Don't-Repeat-Yourself" (DRY) principle.