Remove dependency: error-chain
As the ``error-chain`` crate is hard to understand, and it encourages
using a single massive enum for the entire crate (although this is not
required), a more light weight solution is preferable. As of this
writing, the new error crate ``failure`` seems to be a simpler solution.

``failure`` explicitly lets the user of the crate decide how errors are
defined; the only thing ``failure`` asks is that all errors implement
the ``Fail`` trait. In doing so, code appears to be much simpler. Also,
it is easy to explicitly define all the errors a function may return, no
more and no less.

For example, if a function named ``func`` can possibly raise 2 errors,
``ErrA`` and ``ErrB``, then the function's return type should reflect
that. This is what the ``failure`` crate enables.

This changeset completely removes ``error-chain`` and replaces it with
the ``failure`` crate. Modules and tests have also been updated.