core: message type serialization/deserialization
This is done to make serialization and deserialization more ergonomic for all
message types, as well as to reduce this crate's dependency on the tokio crate.

The initial implementation was built using types provided by the tokio crate.
However, since tokio technically isn't required for the actual
serialization/deserialization process, its use falls outside of the scope of
this crate.

This changeset introduces 2 new traits, ``AsBytes`` and ``FromBytes`` and
implements them for all message types ie ``Message``, ``RequestMessage``,
``ResponseMessage``, and ``NotificationMessage``. Also included in this
changeset are unit tests for all message types using the new traits.

Changes include:

* New trait: AsBytes

  - This trait converts messages into bytes

* New trait: FromBytes

  - This trait creates a message from messagepack bytes

* RpcMessage: switch to using associated types instead of generics

  - This is needed in order to implement the AsBytes and FromBytes traits

* RpcMessage: remove ``value_type_name`` method as it was never being used, and
  a public function ``value_type`` exists providing the same functionality

* All message types implement Clone and PartialEq

  - This is needed to make development of unit tests more ergonomic