Replace *::from() methods with FromMessage trait
Due to needing a conversion method similar to the TryFrom trait, yet
also needing to not depend on nightly rust, ``from`` methods where
explicitly implemented for various concrete types. However, this led to
a lot of copy-pasting of code since the various message types all used
similar conversions.

In order to uphold the Don't-Repeat-Yourself principle, all ``from``
methods were removed and replaced by a new trait ``FromMessage``. The
implementation of this new trait allowed for defining a common
implementation once for all of the message types improving
maintainability. However, this is at the cost of requiring that the
``FromMessage`` trait be imported when needed.