Add config options for rustfmt 0.2.15
As sometimes it is necessary to use an older nightly rust with an older
rustfmt, doing so may screw up the formatting of the project due to
difference of config options between different versions of rustfmt.

In order for the project to be formatted by versions 0.2 and 0.3 of
rustfmt, the rustfmt.toml file is updated to include equivalent options
that result in the same formatting style in both versions.

It is expected, once a version of nightly rust works with a more recent
version of rustfmt, this change will be reverted so that rustfmt.toml
only contains config options for rustfmt v0.3+