FIX: ResponseBuilder::read() doesn't check request
ResponseBuilder receives a request object in order to create a response
to that request. However, it should only build responses with a code
matching the request's method code, and error out otherwise.

As ResponseBuilder::read() didn't check the request to validate the
request's method code, this changeset adds that check in. Also included
is a property test to check that only request's with RequestCode::Read
can be used to build a response with ResponseBuilder::read().